requirements.txt
feedparser>=6.0.0
requests>=2.31.0
google-generativeai>=0.4.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
yfinance>=0.2.0
python-dotenv>=1.0.0
